Change: Extract getNFTs and getTokens balances into `getOnchainBalances.ts`
