+++ v2/libs/portfolio/portfolio.ts
@@ -1,26 +1,14 @@
 import { Provider, JsonRpcProvider } from 'ethers'
-import { Deployless, DeploylessMode, parseErr } from '../deployless/deployless'
-import { AccountOp, callToTuple } from '../accountOp/accountOp'
-import { getAccountDeployParams } from '../account/account'
+import { Deployless } from '../deployless/deployless'
+import { AccountOp } from '../accountOp/accountOp'
 import { Account } from '../../interfaces/account'
 import { NetworkDescriptor } from '../../interfaces/networkDescriptor'
 import { nftOracle, balanceOracle } from './multiOracle.json'
 import batcher from './batcher'
 import { geckoRequestBatcher, geckoResponseIdentifier } from './gecko'
 import { flattenResults, paginate } from './pagination'
-import { TokenResult, Collectable, Price } from './interfaces'
-
-type LimitsOptions = {
-  erc20: number
-  erc721: number
-  erc721TokensInput: number
-  erc721Tokens: number
-}
-
-type Limits = {
-  deploylessProxyMode: LimitsOptions
-  deploylessStateOverrideMode: LimitsOptions
-}
+import { TokenResult, Price, Limits, LimitsOptions } from './interfaces'
+import { getNFTs, getTokens } from './getOnchainBalances'
 
 const LIMITS: Limits = {
   // we have to be conservative with erc721Tokens because if we pass 30x20 (worst case) tokenIds, that's 30x20 extra words which is 19kb
@@ -34,8 +22,6 @@ const LIMITS: Limits = {
     erc721Tokens: 100
   }
 }
-// 0x00..01 is the address from which simulation signatures are valid
-const DEPLOYLESS_SIMULATION_FROM = '0x0000000000000000000000000000000000000001'
 
 export interface GetOptionsSimulation {
   accountOps: AccountOp[]
@@ -204,149 +190,5 @@ export class Portfolio {
         return cur
       }, {})
     }
-  }
-}
-
-const handleSimulationError = (error: string, beforeNonce: bigint, afterNonce: bigint) => {
-  if (error !== '0x') throw new SimulationError(parseErr(error) || error, beforeNonce, afterNonce)
-  if (afterNonce === 0n)
-    throw new SimulationError('unknown error: simulation reverted', beforeNonce, afterNonce)
-  if (afterNonce < beforeNonce)
-    throw new SimulationError('lower "after" nonce', beforeNonce, afterNonce)
-}
-
-async function getNFTs(
-  deployless: Deployless,
-  opts: Partial<GetOptions>,
-  accountAddr: string,
-  tokenAddrs: [string, any][],
-  limits: LimitsOptions
-): Promise<[number, TokenResult][]> {
-  const deploylessOpts = { blockTag: opts.blockTag, from: DEPLOYLESS_SIMULATION_FROM }
-  const mapToken = (token: any) => {
-    return {
-      name: token.name,
-      symbol: token.symbol,
-      amount: BigInt(token.nfts.length),
-      decimals: 1,
-      collectables: [...(token.nfts as any[])].map(
-        (token: any) => ({ id: token.id, url: token.uri } as Collectable)
-      )
-    } as TokenResult
-  }
-
-  if (!opts.simulation) {
-    const collections = (
-      await deployless.call(
-        'getAllNFTs',
-        [
-          accountAddr,
-          tokenAddrs.map(([address]) => address),
-          tokenAddrs.map(([_, x]) =>
-            x.enumerable ? [] : x.tokens.slice(0, limits.erc721TokensInput)
-          ),
-          limits.erc721Tokens
-        ],
-        deploylessOpts
-      )
-    )[0]
-
-    return collections.map((token: any) => [token.error, mapToken(token)])
-  }
-
-  const { accountOps, account } = opts.simulation
-  const [factory, factoryCalldata] = getAccountDeployParams(account)
-
-  const [before, after, simulationErr] = await deployless.call(
-    'simulateAndGetAllNFTs',
-    [
-      accountAddr,
-      tokenAddrs.map(([address]) => address),
-      tokenAddrs.map(([_, x]) => (x.enumerable ? [] : x.tokens.slice(0, limits.erc721TokensInput))),
-      limits.erc721Tokens,
-
-      factory,
-      factoryCalldata,
-      accountOps.map(({ nonce, calls, signature }) => [nonce, calls.map(callToTuple), signature])
-    ],
-    deploylessOpts
-  )
-
-  const beforeNonce = before[1]
-  const afterNonce = after[1]
-  handleSimulationError(simulationErr, beforeNonce, afterNonce)
-
-  // no simulation was performed if the nonce is the same
-  const postSimulationAmounts = (after[1] === before[1] ? before[0] : after[0]).map(mapToken)
-
-  return before[0].map((token: any, i: number) => [
-    token.error,
-    { ...mapToken(token), amountPostSimulation: postSimulationAmounts[i].amount }
-  ])
-}
-
-async function getTokens(
-  network: NetworkDescriptor,
-  deployless: Deployless,
-  opts: Partial<GetOptions>,
-  accountAddr: string,
-  tokenAddrs: string[]
-): Promise<[number, TokenResult][]> {
-  const mapToken = (token: any, address: string) =>
-    ({
-      amount: token.amount,
-      decimals: new Number(token.decimals),
-      symbol:
-        address === '0x0000000000000000000000000000000000000000'
-          ? network.nativeAssetSymbol
-          : token.symbol,
-      address
-    } as TokenResult)
-  const deploylessOpts = { blockTag: opts.blockTag, from: DEPLOYLESS_SIMULATION_FROM }
-  if (!opts.simulation) {
-    const [results] = await deployless.call(
-      'getBalances',
-      [accountAddr, tokenAddrs],
-      deploylessOpts
-    )
-
-    return results.map((token: any, i: number) => [token.error, mapToken(token, tokenAddrs[i])])
-  }
-  const { accountOps, account } = opts.simulation
-  const [factory, factoryCalldata] = getAccountDeployParams(account)
-  const [before, after, simulationErr] = await deployless.call(
-    'simulateAndGetBalances',
-    [
-      accountAddr,
-      tokenAddrs,
-      factory,
-      factoryCalldata,
-      accountOps.map(({ nonce, calls, signature }) => [nonce, calls.map(callToTuple), signature])
-    ],
-    deploylessOpts
-  )
-
-  const beforeNonce = before[1]
-  const afterNonce = after[1]
-  handleSimulationError(simulationErr, beforeNonce, afterNonce)
-
-  // no simulation was performed if the nonce is the same
-  const postSimulationAmounts = afterNonce === beforeNonce ? before[0] : after[0]
-
-  return before[0].map((token: any, i: number) => [
-    token.error,
-    { ...mapToken(token, tokenAddrs[i]), amountPostSimulation: postSimulationAmounts[i].amount }
-  ])
-}
-
-class SimulationError extends Error {
-  public simulationErrorMsg: string
-  public beforeNonce: bigint
-  public afterNonce: bigint
-  constructor(message: string, beforeNonce: bigint, afterNonce: bigint) {
-    super(`simulation error: ${message}`)
-    this.simulationErrorMsg = message
-    this.beforeNonce = beforeNonce
-    this.afterNonce = afterNonce
   }
 }
